Fix home page metadata in client entry
